project structuring + beginning implementation of map
