--- Response.py
@@ -1,1 +1,0 @@
-Hey Big Boi
