+++ map/map.py
@@ -1,0 +1,21 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sat Jul 13 06:47:02 2019
+
+@author: Jamesrulz
+"""
+
+''' Start with territory types then define the map - as we're only using a standard map it probably makes sense to just hardcode
+in the locations and their adjacencies'''
+
+class Map:
+    ''' '''
+    #TODO
+    
+class SupplyCenterMap:
+    ''' '''
+    #TODO
+    
+class OwnershipMap:
+    ''' '''
+    #TODO
